requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
Pillow==10.2.0
numpy==1.26.3
trimesh==4.3.0
pythreejs==2.4.2
python-slugify==8.0.4